+++ CHANGELOG.md
@@ -33,6 +33,7 @@ Types of changes: `Added`, `Changed`, `Deprecate`, `Removed`, `Fixed`, `Secruity
 - Added facade Convert with function for converting a boolean into an integer
 - Added command to create a new model
 - Added web cli
+- Added all CRUD actions for user with UserController
 
 ### Changed
 - Improved the autoloader to work for the new folder structure of version 2